+++ toolbox/library/src/OSQP.cpp
@@ -14,10 +14,9 @@
 #include <BlockFactory/Core/Parameters.h>
 #include <BlockFactory/Core/Signal.h>
 #include <Eigen/Core>
-#include <OsqpEigen/OsqpEigen.h>
 #include <OsqpEigen/Constants.hpp>
+#include <OsqpEigen/OsqpEigen.h>
 #include <osqp/util.h>
-
 
 #include <ostream>
 #include <tuple>
@@ -27,8 +26,6 @@
 #endif
 
 using namespace blockfactory::core;
-
-const unsigned MaxIterations = 100;
 
 #define WBT_OSQP_INF 1e8
 
@@ -44,6 +41,9 @@ enum ParamIndex
     UseUb,
     ComputeObjVal,
     StopWhenFails,
+    AdaptiveRho,
+    Polish,
+    MaxIterations,
 };
 
 enum InputIndex
@@ -83,6 +83,9 @@ public:
     bool useUb;
     bool computeObjVal;
     bool stopWhenFails;
+    bool adaptiveRho;
+    bool polish;
+    unsigned maxIterations;
     // Buffers
     size_t numberOfVariables;
     size_t numberOfTotalConstraints;
@@ -110,7 +113,7 @@ wbt::block::OSQP::~OSQP() = default;
 
 unsigned wbt::block::OSQP::numberOfParameters()
 {
-    return Block::numberOfParameters() + 6;
+    return Block::numberOfParameters() + 9;
 }
 
 bool wbt::block::OSQP::parseParameters(BlockInformation* blockInfo)
@@ -122,6 +125,9 @@ bool wbt::block::OSQP::parseParameters(BlockInformation* blockInfo)
         {ParameterType::BOOL, ParamIndex::UseUb, 1, 1, "UseUb"},
         {ParameterType::BOOL, ParamIndex::ComputeObjVal, 1, 1, "ComputeObjVal"},
         {ParameterType::BOOL, ParamIndex::StopWhenFails, 1, 1, "StopWhenFails"},
+        {ParameterType::BOOL, ParamIndex::AdaptiveRho, 1, 1, "AdaptiveRho"},
+        {ParameterType::BOOL, ParamIndex::Polish, 1, 1, "Polish"},
+        {ParameterType::INT, ParamIndex::MaxIterations, 1, 1, "MaxIterations"},
     };
 
     for (const auto& md : metadata) {
@@ -238,7 +244,7 @@ bool wbt::block::OSQP::configureSizeAndPorts(BlockInformation* blockInfo)
     return true;
 }
 
-bool wbt::block::OSQP::solverInitialization(const BlockInformation*blockInfo)
+bool wbt::block::OSQP::solverInitialization(const BlockInformation* blockInfo)
 {
     // Check the hessian size
     const auto size_H = blockInfo->getInputPortMatrixSize(InputIndex::Hessian);
@@ -284,10 +290,14 @@ bool wbt::block::OSQP::solverInitialization(const BlockInformation*blockInfo)
         // Check the constraints' bound size
         bool ok = true;
         if (pImpl->useLbA) {
-            ok = ok && (blockInfo->getInputPortWidth(InputIndex_lbA) == pImpl->numberOfProperConstraints);
+            ok = ok
+                 && (blockInfo->getInputPortWidth(InputIndex_lbA)
+                     == pImpl->numberOfProperConstraints);
         }
         if (pImpl->useUbA) {
-            ok = ok && (blockInfo->getInputPortWidth(InputIndex_ubA) == pImpl->numberOfProperConstraints);
+            ok = ok
+                 && (blockInfo->getInputPortWidth(InputIndex_ubA)
+                     == pImpl->numberOfProperConstraints);
         }
         if (!ok) {
             bfError << "Sizes of constraints' bounds do not match with the number of constraints.";
@@ -311,20 +321,30 @@ bool wbt::block::OSQP::solverInitialization(const BlockInformation*blockInfo)
     }
 
     // Resize buffers
-    pImpl->totalConstraintsLowerBounds = Eigen::VectorXd::Constant(pImpl->numberOfTotalConstraints, -WBT_OSQP_INF);
-    pImpl->totalConstraintsUpperBounds = Eigen::VectorXd::Constant(pImpl->numberOfTotalConstraints, WBT_OSQP_INF);
-    pImpl->totalConstraintsMatrix = Eigen::MatrixXd(pImpl->numberOfTotalConstraints, pImpl->numberOfVariables);
+    pImpl->totalConstraintsLowerBounds =
+        Eigen::VectorXd::Constant(pImpl->numberOfTotalConstraints, -WBT_OSQP_INF);
+    pImpl->totalConstraintsUpperBounds =
+        Eigen::VectorXd::Constant(pImpl->numberOfTotalConstraints, WBT_OSQP_INF);
+    pImpl->totalConstraintsMatrix =
+        Eigen::MatrixXd(pImpl->numberOfTotalConstraints, pImpl->numberOfVariables);
 
     // Setup options
     pImpl->sqSolver->settings()->setVerbosity(false);
 
-    // Set adaptive_rho to false to avoid converge problems
+    // Setup adaptive_rho
+    // Note that enabling adaptive_rho could cause converge problems
     // See https://github.com/oxfordcontrol/osqp/issues/151
-    pImpl->sqSolver->settings()->setAdaptiveRho(false);
+    pImpl->sqSolver->settings()->setAdaptiveRho(pImpl->adaptiveRho);
 
     // Set warm start to true to permit to just update the values
     // of hessians, gradient and constraints
     pImpl->sqSolver->settings()->setWarmStart(true);
+
+    // Setup the polishing option
+    pImpl->sqSolver->settings()->setPolish(pImpl->polish);
+
+    // Setup the maximum number of iterations
+    pImpl->sqSolver->settings()->setMaxIteration(pImpl->maxIterations);
 
     return true;
 }
@@ -338,6 +358,8 @@ bool wbt::block::OSQP::initialize(BlockInformation* blockInfo)
     // PARAMETERS
     // ==========
 
+    int maxIter = -1;
+
     if (!OSQP::parseParameters(blockInfo)) {
         bfError << "Failed to parse parameters.";
         return false;
@@ -350,9 +372,21 @@ bool wbt::block::OSQP::initialize(BlockInformation* blockInfo)
     ok = ok && m_parameters.getParameter("UseUb", pImpl->useUb);
     ok = ok && m_parameters.getParameter("ComputeObjVal", pImpl->computeObjVal);
     ok = ok && m_parameters.getParameter("StopWhenFails", pImpl->stopWhenFails);
+    ok = ok && m_parameters.getParameter("AdaptiveRho", pImpl->adaptiveRho);
+    ok = ok && m_parameters.getParameter("Polish", pImpl->polish);
+    ok = ok && m_parameters.getParameter("MaxIterations", maxIter);
 
     if (!ok) {
         bfError << "Failed to get parameters after their parsing.";
+        return false;
+    }
+
+    // Check if the maximum number of Iterations read correctly from Simulink GUI
+    if (maxIter > 0) {
+        pImpl->maxIterations = maxIter;
+    }
+    else {
+        bfError << "Failed to set the maximum number of iterations.";
         return false;
     }
 
@@ -362,7 +396,6 @@ bool wbt::block::OSQP::initialize(BlockInformation* blockInfo)
         return false;
     }
 
-
     return true;
 }
 
@@ -371,7 +404,6 @@ bool wbt::block::OSQP::initializeInitialConditions(const BlockInformation* /*blo
     pImpl->sqSolver->clearSolver();
     return true;
 }
-
 
 bool wbt::block::OSQP::output(const BlockInformation* blockInfo)
 {
@@ -388,7 +420,8 @@ bool wbt::block::OSQP::output(const BlockInformation* blockInfo)
         blockInfo->getInputPortMatrixSize(InputIndex::Hessian).rows,
         blockInfo->getInputPortMatrixSize(InputIndex::Hessian).cols);
 
-    pImpl->sqSolver->data()->setNumberOfVariables(blockInfo->getInputPortMatrixSize(InputIndex::Hessian).rows);
+    pImpl->sqSolver->data()->setNumberOfVariables(
+        blockInfo->getInputPortMatrixSize(InputIndex::Hessian).rows);
 
     Eigen::Map<Eigen::VectorXd> gradient_colMajor(
         const_cast<double*>(gradientSignal->getBuffer<double>()),
@@ -449,8 +482,8 @@ bool wbt::block::OSQP::output(const BlockInformation* blockInfo)
                 bfError << "Signal for lb is not valid.";
                 return false;
             }
-            Eigen::Map<const Eigen::VectorXd> variableConstraintsLowerBoundsMap(lbSignal->getBuffer<double>(),
-                                                                                lbSignal->getWidth());
+            Eigen::Map<const Eigen::VectorXd> variableConstraintsLowerBoundsMap(
+                lbSignal->getBuffer<double>(), lbSignal->getWidth());
             pImpl->variableConstraintsLowerBounds = variableConstraintsLowerBoundsMap;
         }
 
@@ -460,48 +493,56 @@ bool wbt::block::OSQP::output(const BlockInformation* blockInfo)
                 bfError << "Signal for ub is not valid.";
                 return false;
             }
-            Eigen::Map<const Eigen::VectorXd> variableConstraintsUpperBoundsMap(ubSignal->getBuffer<double>(),
-                                                                                ubSignal->getWidth());
+            Eigen::Map<const Eigen::VectorXd> variableConstraintsUpperBoundsMap(
+                ubSignal->getBuffer<double>(), ubSignal->getWidth());
             pImpl->variableConstraintsUpperBounds = variableConstraintsUpperBoundsMap;
         }
     }
 
-
     // Let's build the actual constraints passed to OSQP
-    if(pImpl->numberOfTotalConstraints != 0) {
+    if (pImpl->numberOfTotalConstraints != 0) {
         if (pImpl->useLbA) {
-            pImpl->totalConstraintsLowerBounds.head(pImpl->numberOfProperConstraints) = pImpl->properConstraintsLowerBounds;
+            pImpl->totalConstraintsLowerBounds.head(pImpl->numberOfProperConstraints) =
+                pImpl->properConstraintsLowerBounds;
         }
 
         if (pImpl->useUbA) {
-            pImpl->totalConstraintsUpperBounds.head(pImpl->numberOfProperConstraints) = pImpl->properConstraintsUpperBounds;
+            pImpl->totalConstraintsUpperBounds.head(pImpl->numberOfProperConstraints) =
+                pImpl->properConstraintsUpperBounds;
         }
 
-
         if (pImpl->useLb) {
-            pImpl->totalConstraintsLowerBounds.tail(pImpl->numberOfVariables) = pImpl->variableConstraintsLowerBounds;
+            pImpl->totalConstraintsLowerBounds.tail(pImpl->numberOfVariables) =
+                pImpl->variableConstraintsLowerBounds;
         }
 
         if (pImpl->useUb) {
-            pImpl->totalConstraintsUpperBounds.tail(pImpl->numberOfVariables) = pImpl->variableConstraintsUpperBounds;
+            pImpl->totalConstraintsUpperBounds.tail(pImpl->numberOfVariables) =
+                pImpl->variableConstraintsUpperBounds;
         }
 
         if (pImpl->useLbA || pImpl->useUbA) {
-            pImpl->totalConstraintsMatrix.topRows(pImpl->numberOfProperConstraints) = pImpl->properConstraintMatrix;
+            pImpl->totalConstraintsMatrix.topRows(pImpl->numberOfProperConstraints) =
+                pImpl->properConstraintMatrix;
         }
 
         if (pImpl->useLb || pImpl->useUb) {
-            pImpl->totalConstraintsMatrix.bottomRows(pImpl->numberOfVariables) = Eigen::MatrixXd::Identity(pImpl->numberOfVariables, pImpl->numberOfVariables);
+            pImpl->totalConstraintsMatrix.bottomRows(pImpl->numberOfVariables) =
+                Eigen::MatrixXd::Identity(pImpl->numberOfVariables, pImpl->numberOfVariables);
         }
 
         if (!pImpl->sqSolver->isInitialized()) {
             pImpl->totalConstraintsMatrixSparse = pImpl->totalConstraintsMatrix.sparseView();
-            pImpl->sqSolver->data()->setLinearConstraintsMatrix(pImpl->totalConstraintsMatrixSparse);
-            pImpl->sqSolver->data()->setBounds(pImpl->totalConstraintsLowerBounds, pImpl->totalConstraintsUpperBounds);
-        } else {
+            pImpl->sqSolver->data()->setLinearConstraintsMatrix(
+                pImpl->totalConstraintsMatrixSparse);
+            pImpl->sqSolver->data()->setBounds(pImpl->totalConstraintsLowerBounds,
+                                               pImpl->totalConstraintsUpperBounds);
+        }
+        else {
             pImpl->totalConstraintsMatrixSparse = pImpl->totalConstraintsMatrix.sparseView();
             pImpl->sqSolver->updateLinearConstraintsMatrix(pImpl->totalConstraintsMatrixSparse);
-            pImpl->sqSolver->updateBounds(pImpl->totalConstraintsLowerBounds, pImpl->totalConstraintsUpperBounds);
+            pImpl->sqSolver->updateBounds(pImpl->totalConstraintsLowerBounds,
+                                          pImpl->totalConstraintsUpperBounds);
         }
     }
     // OUTPUTS
@@ -515,7 +556,6 @@ bool wbt::block::OSQP::output(const BlockInformation* blockInfo)
     Eigen::Map<Eigen::VectorXd> solution_colMajor(
         const_cast<double*>(solutionSignal->getBuffer<double>()),
         blockInfo->getOutputPortWidth(OutputIndex::PrimalSolution));
-
 
     OutputSignalPtr statusSignal = blockInfo->getOutputPortSignal(OutputIndex::Status);
     if (!statusSignal) {
@@ -532,35 +572,34 @@ bool wbt::block::OSQP::output(const BlockInformation* blockInfo)
     if (!pImpl->sqSolver->isInitialized()) {
         pImpl->sqSolver->data()->setHessianMatrix(hessian_sparse);
         pImpl->sqSolver->data()->setGradient(gradient_colMajor);
-    } else {
+    }
+    else {
         pImpl->sqSolver->updateHessianMatrix(hessian_sparse);
         pImpl->sqSolver->updateGradient(gradient_colMajor);
     }
 
-
     // Solve
-    if(!pImpl->sqSolver->isInitialized()) {
-        if(!pImpl->sqSolver->initSolver()) {
+    if (!pImpl->sqSolver->isInitialized()) {
+        if (!pImpl->sqSolver->initSolver()) {
             bfError << "OSQP: initSolver() failed.";
             return false;
         }
     }
 
-    bool solveReturnVal = pImpl->sqSolver->solve();
+    OsqpEigen::ErrorExitFlag solveReturnVal = pImpl->sqSolver->solveProblem();
 
-    if (pImpl->stopWhenFails && !solveReturnVal) {
-        bfError << "OSQP: solve() failed.";
+    if (pImpl->stopWhenFails && solveReturnVal != OsqpEigen::ErrorExitFlag::NoError) {
+        bfError << "OSQP: solveProblem() failed.";
         return false;
     }
 
     // Get outputs
-    if (solveReturnVal) {
+    if (solveReturnVal == OsqpEigen::ErrorExitFlag::NoError) {
         solution_colMajor = pImpl->sqSolver->getSolution();
     }
 
     // Set status
-    double status = solveReturnVal ? 0 : 1;
-    if (!statusSignal->set(0, status)) {
+    if (!statusSignal->set(0, double(solveReturnVal))) {
         bfError << "Failed to set status signal.";
         return false;
     }